+++ src/LASRstages/localmaximum.cpp
@@ -105,7 +105,6 @@ bool LASRlocalmaximum::process(LAS*& las)
   progress->reset();
   progress->set_total(las->npoints);
   progress->set_prefix("Local maximum");
-  progress->set_ncpu(ncpu);
 
   // Local maximum algorithm
   double hws = ws/2;
@@ -199,12 +198,13 @@ bool LASRlocalmaximum::process(LAS*& las)
 bool LASRlocalmaximum::write()
 {
   if (ofile.empty()) return true;
-  if (lm.size() == 0) return true;
 
   int dupfid= 0;
   progress->reset();
   progress->set_total(lm.size());
   progress->set_prefix("Write local maxima on disk");
+
+  if (lm.size() == 0) return true;
 
   for (const auto& p : lm)
   {